The design of "scholarship management system" adopts php, one of the most popular network programming languages, and the database adopts mysql, which improves the storage security of data. The overall framework of the system adopts content relational structure, which is conducive to the interaction between the system and users. "scholarship  management system" The function design basically meets the needs of the practical application of higher learning institutions and the Liberian society
The design of "scholarship management system" adopts php, one of the most popular network programming languages, and the database adopts mysql, which improves the storage security of data. The overall framework of the system adopts content relational structure, which is conducive to the interaction between the system and users. "scholarship  management system" The function design basically meets the needs of the practical application of higher learning institutions and the Liberian society